refactor(gui): split settings_handler.rs into settings_handler/ submodule
The 950-line `settings_handler.rs` bundled 20+ `save_*` methods for
hotkey, layout, storage, theme, and update concerns in one file. Group
them into themed submodules so each concern lives next to its peers:

- `hotkey.rs`  — activation-key recording and `save_hotkey`
- `layout.rs`  — `save_selected_layout`
- `storage.rs` — `save_max_history` / `save_max_storage` and the
  `has_pending_max_*` helpers
- `theme.rs`   — theme, language, and window-opacity persistence
- `update.rs`  — `save_auto_check_enabled` /
  `save_include_prerelease_enabled`

`mod.rs` keeps shared helpers (notifications, `persist_settings_update`,
selection sync, validation helpers) plus autostart / confirm-mode /
hover-preview handlers that don't fit the five themes.

Refs #64
